+++ src/main/java/lost_no_more/lost_no_more_batch/elasticsearch/job/ElasticStoreJobConfig.java
@@ -1,13 +1,7 @@
-package lost_no_more.lost_no_more_batch.elastic_item.job;
+package lost_no_more.lost_no_more_batch.elasticsearch.job;
 
-import lombok.RequiredArgsConstructor;
-import lost_no_more.lost_no_more_batch.elastic_item.dto.ElasticLostItemDto;
-import lost_no_more.lost_no_more_batch.elastic_item.reader.ElasticStoreReader;
-import lost_no_more.lost_no_more_batch.elastic_item.writer.ElasticStoreWriter;
-import lost_no_more.lost_no_more_batch.global.job_listner.JobCompletionListener;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
-import org.springframework.batch.core.configuration.annotation.JobScope;
 import org.springframework.batch.core.job.builder.JobBuilder;
 import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.batch.core.step.builder.StepBuilder;
@@ -15,12 +9,16 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.transaction.PlatformTransactionManager;
 
+import lombok.RequiredArgsConstructor;
+import lost_no_more.lost_no_more_batch.elasticsearch.dto.ElasticLostItemDto;
+import lost_no_more.lost_no_more_batch.elasticsearch.reader.ElasticStoreReader;
+import lost_no_more.lost_no_more_batch.elasticsearch.writer.ElasticStoreWriter;
+
 @Configuration
 @RequiredArgsConstructor
 public class ElasticStoreJobConfig {
 
     private final JobRepository jobRepository;
-    private final JobCompletionListener jobCompletionListener;
     private final ElasticStoreReader reader;
     private final PlatformTransactionManager dataTransactionManager;
     private final ElasticStoreWriter writer;
@@ -34,12 +32,10 @@ public class ElasticStoreJobConfig {
                 .on("*")
                 .end()
                 .end()
-                .listener(jobCompletionListener)
                 .build();
     }
 
     @Bean
-    @JobScope
     public Step dailyElasticStoreStep() {
         return new StepBuilder("dailyElasticStoreStep", jobRepository)
                 .<ElasticLostItemDto, ElasticLostItemDto>chunk(1000, dataTransactionManager)
