+++ src/main/java/lost_no_more/lost_no_more_batch/elasticsearch/repository/ElasticStoreRepository.java
@@ -1,6 +1,6 @@
-package lost_no_more.lost_no_more_batch.elastic_item.repository;
+package lost_no_more.lost_no_more_batch.elasticsearch.repository;
 
-import lost_no_more.lost_no_more_batch.elastic_item.domain.LostItemDocument;
+import lost_no_more.lost_no_more_batch.elasticsearch.domain.LostItemDocument;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 public interface ElasticStoreRepository extends ElasticsearchRepository<LostItemDocument, Long> {
