+++ src/main/java/lost_no_more/lost_no_more_batch/item/domain/Category.java
@@ -11,22 +11,23 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lost_no_more.lost_no_more_batch.global.domain.BaseEntity;
 
 @Getter
 @Table(name = "category",
     indexes = {
-        @Index(name = "idx_category_name", columnList = "category_name", unique = true)
+        @Index(name = "idx_name", columnList = "name", unique = true)
     })
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Category {
+public class Category extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "category_id")
     private Long id;
 
-    @Column(name = "category_name", nullable = false)
+    @Column(name = "name", nullable = false)
     private String name;
 
     @Builder
